deleting qml transform model, moving generate_unique_name into ui_utils.py
