+++ nexus_constructor/ui_utils.py
@@ -2,6 +2,7 @@ import numpy as np
 from PySide2.QtGui import QVector3D
 from PySide2.QtWidgets import QFileDialog
 from nexus_constructor.file_dialog_options import FILE_DIALOG_NATIVE
+import re
 
 
 def file_dialog(is_save, caption, filter):
@@ -54,3 +55,22 @@ def qvector3d_to_numpy_array(input_vector: QVector3D):
     return np.array([input_vector.x(), input_vector.y(), input_vector.z()]).astype(
         float
     )
+
+
+def generate_unique_name(base: str, items: list):
+    """
+    Generates a unique name for a new item using a common base string
+
+    :param base: The generated name will be the base string, followed by a number if required
+    :param items: The named items to avoid generating a matching name with. Each must have a 'name' attribute
+    """
+    regex = f"^{re.escape(base)}\\d*$"
+    similar_names = [item.name for item in items if re.match(regex, item.name)]
+
+    if len(similar_names) == 0 or base not in similar_names:
+        return base
+    if similar_names == [base]:
+        return base + "1"
+    # find the highest number in use, and go one higher
+    tailing_numbers = [int(name[len(base) :]) for name in similar_names if name != base]
+    return base + str(max(tailing_numbers) + 1)
